pyproject.toml
[project]
name = "weatheragent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.18.0",
    "python-telegram-bot>=22.5",
]
